refactor(routes): use camelCase route params
