+++ app/routes/_protected.settings.contacts.$contactId.tsx
@@ -2,10 +2,10 @@ import { Page } from '~/components/page';
 import { Redirect } from '~/components/redirect';
 import { useContact } from '~/features/contacts/contact-hooks';
 import { SingleContact } from '~/features/settings/contact';
-import type { Route } from './+types/_protected.settings.contacts.$contact_id';
+import type { Route } from './+types/_protected.settings.contacts.$contactId';
 
 export default function ContactsPage({ params }: Route.ComponentProps) {
-  const contactId = params.contact_id;
+  const contactId = params.contactId;
   const contact = useContact(contactId);
 
   if (!contact) {
